tests infrstructure refactor
